Removed original_shape from manipulator return value (it's stored on client-side, so there seems to be no need to return it from the server).  This required adjustments in individual manipulators as well as BaseManipulator result() method.
Generic view in manipulators now returns 3 versions of clipped geometry to client, a kml version (used for client-side display), a wkt version (used for db storage and passing shape back to server for further analysis), and a stripped down version of the shapes coordinates (used for building new targetShape/placemark in draw_tool).  I'm not happy with having to return so many copies of what is essentially the same thing from the server.  I'd like to find a better solution than this.

Added load and save functionality to client and server side.  This functionality still lacks any sort of identification mechanism to determine whether the save action is on a new mpa or one that already exists in the database (i.e. an mpa that was loaded from the db, modified, and then saved back to the db).

Added several templates in mpa/templates, a new file mpa/forms.py, new views in mpa/views, and changes to mpa/urls.py.

New classes were added to the lingcod namespace including, MpaSaver and MpaLoader, and new functions were added to MpaCreator and DrawTool mostly related to loading and saving.

There are some problems with the way I'm handling the load and save forms (such as separating the submit button from the form), as well as other problems that I'll be addressing soon.

This new functionality can be viewed through mlpa-manipulators (http://localhost:8000/mlpa/mlpa-manipulators/) or through the base app (http://localhost:8000/)